+++ src/rag_kb/graph/__init__.py
@@ -1,0 +1,5 @@
+"""Knowledge Graph Module for RAG KB"""
+
+from .generator import KnowledgeGraphGenerator, graph_generator
+
+__all__ = ['KnowledgeGraphGenerator', 'graph_generator']